% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african_market.R
\docType{class}
\name{african_market-class}
\alias{african_market-class}
\title{S4 Class african_market}
\description{
A class that represents market data with its indexes, shares and Bonds.
}
\section{Slots}{

\describe{
\item{\code{Market_short_name}}{Abbreviated name of the market (character)}

\item{\code{Market_full_name}}{Full name of the market (character)}

\item{\code{Market_url}}{URL of the market (character)}

\item{\code{List}}{List of tickers (character)}

\item{\code{Indexes}}{Index data (data.frame)}

\item{\code{Shares}}{Share data (data.frame)}

\item{\code{Bonds}}{Bond data (data.frame)}
}}

\author{
Koffi Frederic SESSIE

Olabiyi Aurel Géoffroy ODJO
}
