% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african_market.R
\name{african_market}
\alias{african_market}
\title{Constructor for the 'african_market' S4 class}
\usage{
african_market(
  Market_short_name,
  Market_full_name,
  Market_url,
  Market_data_url,
  List,
  ListShares,
  ListIndexes,
  ListBonds,
  Indexes,
  Shares,
  Bonds,
  Ticker_full_name
)
}
\arguments{
\item{Market_short_name}{A character string for the abbreviated name of the market.}

\item{Market_full_name}{A character string for the full name of the market.}

\item{Market_url}{A character string for the URL of the market.}

\item{List}{A character vector of the market's tickers.}

\item{Indexes}{A data frame containing data for the market's indexes.}

\item{Shares}{A data frame containing data for the market's shares.}

\item{Bonds}{A data frame containing data for the market's bonds.}
}
\value{
A new S4 object of class 'african_market'.
}
\description{
Creates an object of the S4 class 'african_market' to represent data for an African stock market.
}
