% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-bysector.R
\name{BRVM_bySector}
\alias{BRVM_bySector}
\alias{BRVM_bySector,character-method}
\title{BRVM By Sector}
\usage{
BRVM_bySector(.sectors = "ALL")

\S4method{BRVM_bySector}{character}(.sectors = "ALL")
}
\arguments{
\item{.sectors}{A vector of sectors you wish to have returned.}
}
\value{
A tibble
}
\description{
This function will take in the name of sector(s) and returns data for companies belonging to that/those sector(s)
}
\details{
This function will take in a vector of sectors provided by the end
user and return a tibble.
}
\examples{
\donttest{
library(stringr)
library(rlang)
library(xml2)
library(rvest)
library(dplyr)

sectors <- c("Industriels","Administration","PUblic service", "AUtr", "SeRVICES FINaNCIERS",
            "distribution", 25)
BRVM_bySector(.sectors = sectors)
BRVM_bySector(.sectors = "All")
BRVM_bySector(.sectors = c("Other", "Agriculture"))
}

}
\seealso{
Other Data Retrieval: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()}
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO

Oudouss Diakite Abdoul

Steven P. Sanderson II, MPH
}
\concept{Data Retrieval}
\concept{Sector}
