% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-cap.R
\name{BRVM_cap}
\alias{BRVM_cap}
\title{BRVM Capitalization}
\usage{
BRVM_cap()
}
\value{
A tibble
}
\description{
It receives no argument and returns informations about BRVM capitalization
}
\examples{

BRVM_cap()


}
\seealso{
\url{https://www.brvm.org/en/capitalisations/0}

Other Data Retrieval: 
\code{\link{BRVM_bySector}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()}

Other BRVM: 
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO
}
\concept{BRVM}
\concept{Data Retrieval}
