% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-company-cap.R
\name{BRVM_company_cap}
\alias{BRVM_company_cap}
\title{BRVM companies capitalization value - To get the capitalization value data of all companies listed on BRVM stock exchange}
\usage{
BRVM_company_cap()
}
\value{
A tibble
}
\description{
This function will get the capitalization value data of all companies listed on the BVRM exchange.
The function takes no parameter
}
\examples{

BRVM_company_cap()


}
\seealso{
Other BRVM: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE
}
\concept{BRVM}
