% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-company-info.R
\name{BRVM_company_info}
\alias{BRVM_company_info}
\alias{BRVM_company_info,character-method}
\title{BRVM Get info about a ticker (Beta, RSI, Closing, Valorisation, etc.)}
\usage{
BRVM_company_info(ticker = "ALL")

\S4method{BRVM_company_info}{character}(ticker = "ALL")
}
\arguments{
\item{ticker}{The ticker of a company or index listed on the BRVM}
}
\value{
A tibble containing indicators and values for the given ticker(s)
}
\description{
Retrieves detailed financial information (RSI, Beta, closing price, etc.)
for a company or index listed on the BRVM stock exchange.
The input is automatically converted to upper case.
}
\examples{
\donttest{
BRVM_company_info("BOAS")
BRVM_company_info("BoaM")
BRVM_company_info("BRVMAG")
BRVM_company_info("ALL INDEXES")
BRVM_company_info("ALL SHARES")
BRVM_company_info("ALL")
}

}
\seealso{
\url{https://www.sikafinance.com}
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO
}
