% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-company-rank.R
\name{BRVM_company_rank}
\alias{BRVM_company_rank}
\alias{BRVM_company_rank,missing-method}
\title{BRVM Company Ranking by Daily Performance}
\usage{
BRVM_company_rank(object)

\S4method{BRVM_company_rank}{missing}(object)
}
\arguments{
\item{object}{No need to give an argument.}
}
\value{
A tibble containing the following columns:
\itemize{
\item \code{ticker} – The stock ticker symbol.
\item \code{company_name} – The name of the company.
\item \code{percent_change} – The daily percentage change.
\item \code{rank} – The company's position in the ranking (1 = highest increase).
}

"tbl_df"     "tbl"        "data.frame"
}
\description{
Retrieves daily trading data for all companies listed on the
Bourse Régionale des Valeurs Mobilières (BRVM) and ranks them
according to their percentage change in share price.

This function returns companies rank from the BRVM Bourse exchange according to their daily change (variation).
}
\details{
The function scrapes the official BRVM website to collect trading
information such as traded volume, previous price, opening price,
closing price, and daily percent change.
After cleaning and converting numeric fields, companies are ranked
in descending order of percentage change, making it easy to identify
the day's top gainers and losers.

This function will get the rank of the companies listed on the BVRM exchange through the Rich Bourse site.
The function takes no parameter
}
\examples{
\donttest{
library(dplyr)
library(rvest)
BRVM_company_rank()
comp.rank <- BRVM_company_rank()
comp.rank<-comp.rank\%>\%
dplyr::arrange(desc(percent_change))
comp.rank
}
}
\seealso{
\url{https://www.brvm.org/en/cours-actions/0}

Other Data Retrieval: 
\code{\link{BRVM_bySector}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()}

Other BRVM: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO

Oudouss Diakite Abdoul

Steven P. Sanderson II, MPH
}
\concept{BRVM}
\concept{Data Retrieval}
