% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-company_url.R
\name{BRVM_company_url}
\alias{BRVM_company_url}
\alias{BRVM_company_url,character-method}
\title{BRVM company url}
\usage{
BRVM_company_url(ticker = "ALL")

\S4method{BRVM_company_url}{character}(ticker = "ALL")
}
\arguments{
\item{ticker}{is the name of the Ticker}
}
\value{
URLs
}
\description{
It receives the ticker name and return an URL
}
\examples{
\dontrun{
    BRVM_company_url("ALL")
}

}
\seealso{
Other BRVM: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO
}
\concept{BRVM}
