% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-rank.R
\name{BRVM_rank}
\alias{BRVM_rank}
\alias{BRVM_rank,character,numeric-method}
\title{BRVM Capitalization Ranking}
\usage{
BRVM_rank(select = "Top", N = 10)

\S4method{BRVM_rank}{character,numeric}(select = "Top", N = 10)
}
\arguments{
\item{select}{A character string indicating whether to return the
"Top" or "Bottom" ranked companies. Default is "Top".}

\item{N}{An integer specifying the number of companies to return.
Default is 10.}
}
\value{
A tibble containing company names, market capitalization, and related
information from the BRVM website.
}
\description{
Retrieves market capitalization data of companies listed on the
Bourse Régionale des Valeurs Mobilières (BRVM) and ranks them
in descending order. Users can choose to extract either the top
or bottom \emph{n} companies by their global capitalization.
}
\examples{
\donttest{
# Top 10 companies by capitalization
BRVM_rank("Top", 10)

# Bottom 5 companies by capitalization
BRVM_rank("Bottom", 5)
}

}
\seealso{
\url{https://www.brvm.org/en/capitalisations/0}

Other Data Retrieval: 
\code{\link{BRVM_bySector}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()}

Other BRVM: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO
}
\concept{BRVM}
\concept{Data Retrieval}
