% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-ticker.R
\name{BRVM_tickers}
\alias{BRVM_tickers}
\alias{BRVM_tickers,missing-method}
\title{Get BRVM Tickers}
\usage{
BRVM_tickers(object)

\S4method{BRVM_tickers}{missing}(object)
}
\arguments{
\item{object}{An object of class \code{missing}, indicating this method is a constructor that does not require an input object.}
}
\value{
A new S4 object of class \code{market_description} populated with ticker data from the BRVM.
}
\description{
A method that retrieves and processes ticker data for indexes and shares from the BRVM website.
Retrieves information about companies listed on the Bourse Régionale des Valeurs Mobilières (BRVM).
This function returns an S4 object containing the list of tickers, detailed shares information,
and (optionally) BRVM indexes.
}
\details{
The BRVM (Bourse Régionale des Valeurs Mobilières) is a regional stock exchange serving the
West African Economic and Monetary Union (WAEMU / UEMOA).
}
\examples{
\dontrun{
BRVM_tickers()
ticks <- BRVM_tickers()
dput(ticks$Ticker) ## Returns the name of all tickers
}

\donttest{
# Retrieve BRVM tickers
library(rvest)
library(httr2)
library(tidyr)
brvm_tickers <- BRVM_tickers()

# Display shares
brvm_tickers$Shares

# List of tickers
brvm_tickers$List

# Print object
brvm_tickers
}

}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO

Oudouss Diakite Abdoul

Steven P. Sanderson II, MPH
}
