% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african-market.R
\name{$,african_market-method}
\alias{$,african_market-method}
\title{Method for accessing slots with $}
\usage{
\S4method{$}{african_market}(x, name)
}
\arguments{
\item{x}{An object of class african_market.}

\item{name}{The name of the slot.}
}
\description{
Allows access to slots using the $ notation, similar to S3.
}
