% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_markets.R
\name{CREATE_ALL_MARKETS}
\alias{CREATE_ALL_MARKETS}
\title{Create and return all available markets}
\usage{
CREATE_ALL_MARKETS(object)
}
\arguments{
\item{object}{ignored (use the "missing" method)}
}
\value{
A list of objects of type \code{market_data} or derived classes
}
\description{
Create and return all available markets
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Géoffroy ODJO
}
