% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-normality-test.R
\name{normality_test}
\alias{normality_test}
\alias{normality_test,numeric,character-method}
\alias{normality_test,ts,character-method}
\title{Normality test with univariate data}
\usage{
normality_test(x, type.test = "ALL")

\S4method{normality_test}{numeric,character}(x, type.test = "ALL")

\S4method{normality_test}{ts,character}(x, type.test = "ALL")
}
\arguments{
\item{x}{a numeric vector or time series..}

\item{type.test}{character such as "Anderson-Darling","Shapiro-Wilk","Jarque Bera","Cramer-von Mises","Shapiro-Francia","Lilliefors (Kolmogorov-Smirnov)","Pearson chi-square", "Agostino".}
}
\value{
a number that indicates the P-value of the normality test
}
\description{
Performs different types of normality test.
}
\examples{
\donttest{
library(goftest)
library(tseries)
library(nortest)
library(fBasics)

# one and a half week stock index
# data including a weekend
y <- ts(c(5353.08,5409.24,5315.57,5270.53, 5211.66,NA,NA,5160.80,5172.37,5160.80,5172.37))

normality_test(y ,"Shapiro-Wilk")

# You can test multiple methods simultaneously
normality_test(y,type.test = c("anderson-darling","Shapiro-Wilk","agostino"))

}

}
\seealso{
\code{stationarity_test}

Other Test: 
\code{\link{stationarity_test}()}

Other BRVM: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_sector}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO
}
\concept{BRVM}
\concept{Test}
