% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african-market.R
\name{show,african_market-method}
\alias{show,african_market-method}
\title{Show method for african_market}
\usage{
\S4method{show}{african_market}(object)
}
\arguments{
\item{object}{An object of class african_market.}
}
\description{
Show method for african_market
}
