% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african-market.R
\docType{class}
\name{african_market-class}
\alias{african_market-class}
\title{African Market Class}
\value{
An object of class \code{african_market}.
}
\description{
An S4 class designed to represent and store structured information
about African financial markets, including metadata, trading instruments,
and market data sources.
}
\section{Slots}{

\describe{
\item{\code{Market_short_name}}{A short abbreviation of the market name (e.g. "BRVM").}

\item{\code{Market_full_name}}{The full name of the market (e.g. "Bourse Régionale des Valeurs Mobilières").}

\item{\code{Official_url}}{The official website of the market.}

\item{\code{Market_url}}{A general URL for accessing market information.}

\item{\code{Market_data_url}}{A URL pointing to detailed market data.}

\item{\code{List}}{A string summarizing available lists of instruments.}

\item{\code{ListShares}}{A string or URL referencing the list of shares.}

\item{\code{ListIndexes}}{A string or URL referencing the list of indexes.}

\item{\code{ListBonds}}{A string or URL referencing the list of bonds.}

\item{\code{Indexes}}{A data frame containing index-level information.}

\item{\code{Shares}}{A data frame containing share-level information.}

\item{\code{Bonds}}{A data frame containing bond-level information.}

\item{\code{Ticker_full_name}}{The full name corresponding to a ticker symbol.}
}}

\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO
}
