% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african-market.R
\name{african_market}
\alias{african_market}
\alias{african_market,character,character,character,character,character,character,character,character,character,data.frame,data.frame,data.frame,character-method}
\title{Constructor for the 'african_market' S4 class}
\usage{
african_market(
  Market_short_name,
  Market_full_name,
  Official_url,
  Market_url,
  Market_data_url,
  List,
  ListShares,
  ListIndexes,
  ListBonds,
  Indexes,
  Shares,
  Bonds,
  Ticker_full_name
)

\S4method{african_market}{character,character,character,character,character,character,character,character,character,data.frame,data.frame,data.frame,character}(
  Market_short_name = "",
  Market_full_name = "",
  Official_url = "",
  Market_url = "",
  Market_data_url = "",
  List = "",
  ListShares = "",
  ListIndexes = "",
  ListBonds = "",
  Indexes = data.frame(),
  Shares = data.frame(),
  Bonds = data.frame(),
  Ticker_full_name = ""
)
}
\arguments{
\item{Market_short_name}{A character string for the abbreviated name of the market (e.g. "BRVM").}

\item{Market_full_name}{A character string for the full name of the market (e.g. "Bourse Régionale des Valeurs Mobilières").}

\item{Official_url}{A character string for the official website of the market.}

\item{Market_url}{A character string for the general URL of the market.}

\item{Market_data_url}{A character string for the URL to access detailed market data.}

\item{List}{A character string describing the types of lists available (e.g. "Shares, Indexes, Bonds").}

\item{ListShares}{A character string or URL pointing to the list of shares.}

\item{ListIndexes}{A character string or URL pointing to the list of indexes.}

\item{ListBonds}{A character string or URL pointing to the list of bonds.}

\item{Indexes}{A data frame containing information on market indexes.}

\item{Shares}{A data frame containing information on market shares.}

\item{Bonds}{A data frame containing information on market bonds.}

\item{Ticker_full_name}{A character string for the full name associated with a ticker symbol.}
}
\value{
An object of class \code{african_market}.

A new S4 object of class 'african_market'.
}
\description{
Creates an object of class \code{african_market}, which stores metadata and financial
data related to an African stock market. This includes basic identifiers, official URLs,
and structured data frames for indexes, shares, and bonds.
}
\details{
This constructor initializes an S4 object of class \code{african_market},
which is useful for storing structured data and metadata for African financial markets.
It can serve as the backbone for functions that fetch, analyze, and visualize market data.
}
\seealso{
\code{\link{setClass}}, \code{\link{african_market}}
}
\concept{African Markets}
