% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-all-markets.R
\name{CREATE_ALL_MARKETS}
\alias{CREATE_ALL_MARKETS}
\alias{CREATE_ALL_MARKETS,missing-method}
\title{Create and return all available markets}
\usage{
CREATE_ALL_MARKETS(object)

\S4method{CREATE_ALL_MARKETS}{missing}(object)
}
\arguments{
\item{object}{ignored (use the "missing" method)}
}
\value{
A list of objects of type \code{market_data} or derived classes
}
\description{
Create and return all available markets
}
\author{
Koffi Frederic SESSIE

Olabiyi Aurel Geoffroy ODJO
}
